AI_analysis/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.0.2
langchain-google-genai>=0.0.3

# Language models
google-generativeai>=0.3.0
transformers>=4.36.0
sentence-transformers>=2.2.2

# Vector stores
faiss-cpu>=1.7.4

# Graph-based workflow
langgraph>=0.0.15

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# File handling
numpy>=1.24.0